LPS-65012 Social-activities-web move constant class back and export the package
