+++ modules/apps/collaboration/social/social-activities-web/bnd.bnd
@@ -1,6 +1,7 @@
 Bundle-Name: Liferay Social Activities Web
 Bundle-SymbolicName: com.liferay.social.activities.web
-Bundle-Version: 1.0.4
+Bundle-Version: 1.1.0
+Export-Package: com.liferay.social.activities.web.constants
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Social
 Web-ContextPath: /social-activities-web
