+++ modules/apps/collaboration/social/social-activities-web/src/main/resources/com/liferay/social/activities/web/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
